環境変数の手動切り替え実装準備
