+++ README.rst
@@ -1,3 +1,23 @@
+この proxyUtil は、PyPAC をフォークし、機能を追加した者です。
+＜追加機能＞
+virtualProxyEnviron()
+    このオブジェクトは、PACから取得したプロキシ設定値を環境変数に適用するためのものです。
+    def __init__(self, proxy_auth, url):
+        ＜パラメータ＞
+        proxy_auth = None
+            プロキシ認証。デフォルトは認証なし
+        URL = https://www.riken.jp/
+            このURLと通信するためのPACを自動で取得します。デフォルトは理化学研究所
+    def set_environ(self):
+        環境変数をセットします。
+    def unset_environ(self):
+        環境変数の設定を元に戻します。
+
+以下、およびその他のドキュメントについては、pypac、あるいはPyPAC表記をproxyUtilに読み替えてください。
+フォーク元のPyPACへの影響が最小限となるように勤めています。
+    
+
+
 PyPAC: Proxy auto-config for Python
 ===================================
